requirements.txt
python-dotenv==1.1.1
python_fasthtml==0.12.30
pytz==2025.2
supabase==2.22.0
